refactor project, paths and definitionbuilder classes
extract functionality into classes and rework import structure
