--- dummy/_godot_defs/dynamic/@actions.d.ts
@@ -1,1 +1,0 @@
-declare type Action = never
