--- dummy/_godot_defs/dynamic/@asset_paths.d.ts
@@ -1,11 +1,0 @@
-
-declare type AssetType = {
-  'res://Main.tscn': PackedScene<Node2D>,
-  'res://icon.png': StreamTexture
-}
-
-declare type SceneName =
-  | 'res://Main.tscn'
-
-declare type AssetPath = keyof AssetType;
-
