--- dummy/_godot_defs/dynamic/@groups.d.ts
@@ -1,2 +1,0 @@
-declare type Groups = {
-}
