--- dummy/_godot_defs/dynamic/@scenes.d.ts
@@ -1,1 +1,0 @@
-export const MainTscn: PackedScene<Node2D>;
