+++ generate_library_defs/generate_base.ts
@@ -1,7 +1,7 @@
 import fs from "fs"
 import path from "path"
 
-import { TsGdProjectClass } from "../project/project"
+import { Paths } from "../project"
 
 import { ArrayDefinition } from "./custom_defs/array_def"
 import { DictionaryDefinition } from "./custom_defs/dictionary_def"
@@ -266,9 +266,9 @@ declare class Signal<T extends (...args: any[]) => any = () => void> {
 }
 `
 
-export const buildBase = () => {
+export default function writeBaseDefinitions(paths: Paths) {
   fs.writeFileSync(
-    path.join(TsGdProjectClass.Paths.staticGodotDefsPath, "@base.d.ts"),
+    path.join(paths.staticGodotDefsPath, "@base.d.ts"),
     baseFileContent
   )
 }
