+++ generate_library_defs/generate_gdscript_lib.ts
@@ -6,7 +6,7 @@ import {
   formatJsDoc,
   godotTypeToTsType,
   sanitizeGodotNameForTs,
-} from "./generate_library"
+} from "./generation_utils"
 
 export type GodotXMLMethod = {
   $: {
