+++ project/assets/asset_font.ts
@@ -1,20 +1,20 @@
-import { TsGdProjectClass } from "../project"
+import TsGdProject from "../project"
 
 import { BaseAsset } from "./base_asset"
 
 export class AssetFont extends BaseAsset {
   resPath: string
   fsPath: string
-  project: TsGdProjectClass
+  project: TsGdProject
 
   static extensions() {
     return [".ttf"]
   }
 
-  constructor(path: string, project: TsGdProjectClass) {
+  constructor(path: string, project: TsGdProject) {
     super()
     this.fsPath = path
-    this.resPath = TsGdProjectClass.FsPathToResPath(this.fsPath)
+    this.resPath = project.paths.fsPathToResPath(this.fsPath)
     this.project = project
   }
