+++ project/generate_dynamic_defs/index.ts
@@ -1,0 +1,2 @@
+export * from "./definition_builder"
+export { default } from "./definition_builder"
