+++ project/godot_project_file.ts
@@ -1,7 +1,7 @@
 import fs from "fs"
 
-import { TsGdProjectClass as TsGdProject } from "./project"
 import { parseGodotConfigFile } from "./godot_parser"
+import TsGdProject from "./project"
 
 interface IRawGodotConfig {
   globals: {
@@ -139,7 +139,7 @@ export class GodotProjectFile {
 
     return {
       resPath: mainSceneResPath,
-      fsPath: TsGdProject.ResPathToFsPath(mainSceneResPath),
+      fsPath: this.project.paths.resPathToFsPath(mainSceneResPath),
     }
   }
 }
