+++ project/index.ts
@@ -1,0 +1,3 @@
+export { Paths } from "./paths"
+export { TsGdProject } from "./project"
+export { default } from "./project"
