+++ tests/test.ts
@@ -9,6 +9,7 @@ import { ParseNodeType, parseNode } from "../parse_node"
 import { Scope } from "../scope"
 import { TsGdError, __getErrorsTestOnly } from "../errors"
 import { baseContentForTests } from "../generate_library_defs/generate_base"
+import { Paths } from "../project/paths"
 
 import { createStubSourceFileAsset } from "./stubs"
 
@@ -70,6 +71,7 @@ export const compileTs = (code: string, isAutoload: boolean): ParseNodeType => {
   const sourceFileAsset = createStubSourceFileAsset("Test")
 
   // TODO: Make this less silly.
+  // I suppose we could actually use the dummy project
   const godotFile = parseNode(sourceFile, {
     indent: "",
     sourceFile: sourceFile,
@@ -92,6 +94,8 @@ export const compileTs = (code: string, isAutoload: boolean): ParseNodeType => {
       shouldBuildLibraryDefinitions: () => false,
       validateAutoloads: () => [],
       buildLibraryDefinitions: async () => {},
+      paths: {} as any,
+      definitionBuilder: {} as any,
       mainScene: {
         fsPath: "",
         resPath: "",
