+++ tsconfig.json
@@ -53,8 +53,8 @@
     // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "exclude": [
-    "./_godot_defs/**",
-    "./dummy"
-  ]
+  "exclude": ["./_godot_defs/**", "./dummy"],
+  "ts-node": {
+    "files": true
+  }
 }
